Allow a single lint
Fixing this warning would require passing in a struct (a good idea to
start with), but that doesn't match the style of the rest of the
codebase. In stead of sticking out, let's allow just this command to
have bunch of parameters. We can remove this `allow` later if command
line argument processing ever gets refactored.