Remove accidental UTF-8 "emdash" in payment preview help message.
When previewing a payment the help text seeks to remind the user that the
transaction will not be committed unless the command is re-run with the
"--commit" flag. However, an overzealous code editor appears to have taken the
double hyphen and replaced it with a singular U+2014 (Em Dash) character in
the UTF-8 encoding. This is not helpful as it tends to render as a single
dash in the terminal (if it renders at all).

Replace the character with the proper double hyphen ASCII sequence.