Add 24 word BIP-39 seed phrase support (#119)
* Add suport for 12/24 BIP39 seed phrases

Change the --seed option to take a "type" bip39 or mobile

The "--seed mobile" option will behave exactly like the --seed option
did. It will accept only 12 word phrases from the mobile app, verifying
that the checksum is 0.

The "--seed bip39" option will accept valid 12 or 24 word BIP39 phrases
verifying they have the correct BIP39 checksum.

There are probably a few places I could have factored out some repetitive
code, but I chose to leave it as is so the PR wasn't as big and hopefully
eaiser to see what the changes are.

Note that this does not add path derivation or anything else many people
expect from a BIP39 phrase. It is a first step towards that.