Remove seed_type cmdline parameter everywhere
In an effort to be easier to use and clean up the code, remove the
seed_type bip39|mobile parameter. Rather than making the user be
explicit. The code will now behave this way:

* On create, allow bip39 or '0' checksum for 12 word phrases. 24
word phrases always require bip39 checksum

* On verify, always output a phrase with a bip39 checksum.

A side effect of this is that if a user inputs a hotspot-app phrase
during create, the verify command will output a slightly different
phrase. It's the same wallet secret, though. There's even a unit test
to prove this is true.

I think this matches the behavior of the soon-to-be released
helium wallet app.

Along the way, cleaned up some unused dependencies.