Check OUI addresses when paying
Router has had users `pay` to the address rather than `burn` which can
be confusing. The transaction goes through, but the user doesn't see
their DC balance go up.

Respecting that this command needs to be usable when airgapped, we're
working off the assumption that when a nonce is provided no network
calls need to be done, and most likely the user knows what they're
doing.

Otherwise, we can help prevent unnecessary transactions by checking if
the address being payed is pointed to by an OUI.