QR code export of seed for importing into a mobile app.
Add an `export` command, which asks for a password, then uses that
password to encrypt a copy of the mnemonic seed phrase. Finally,
output a QR code of a JSON blob which contains the encrypted text
along with the nonce and seed needed to decrypt it. The user will
have to supply the original password to decrypt.

The process is roughly like this:
* Generate a secret key using pwhash() set to use the argon2id13
algorithm with MODERATE ops and memlimits. The settings here were
chosen to be usable on mobile devices during decryption.

* Encrypt the mnemonic phrase using libsodium's `secretbox_easy()`
set to use the xsalsa20poly1305 algorithm. This algorithm was
chosen because it is available in the existing sodium npm the mobile
wallets use.

* Output some JSON that includes the ciphertext, the salt, and the
nonce all as base64-encoded strings. The base64 encoding is so
that the resulting JSON contains only strings, making debugging
easier.

* Turn the JSON into a QR code, output that.