Fix asset to hotspot conversion
Asset convertion to a Hotspot was relying on a metadata field that may take over 24 hours to get populated, and is only to be used for non critical purposes.

This PR introduces a KeyToAsset cache that is used to cache asset accounts based on the asset key. The cache is then used for Asset->Hotspot conversion to get the entity key (back)