add helper for going from solana keypair to wrapped helium-lib keypair
The solana_sdk provides a `read_keypair_file` function, but it returns a
Result with a Box<dyn Error>. Rather than recreate all those functions
for a wrapped Keypair, I thought to start with an easy conversion from
one to the other. So you can read the keypair with the solana_sdk, then
very quickly turn it into a helium-lib keypair.

Bring the discussion!!!