Updates for Oracles to transition to helium-lib (#401)
* Move SubDao::dc_key() -> Dao::dc_key()

The Data Credits Key is not SubDao specific.

* Add `dc::burn_delegated` function

* add helper for going from solana keypair to wrapped helium-lib keypair

The solana_sdk provides a `read_keypair_file` function, but it returns a
Result with a Box<dyn Error>. Rather than recreate all those functions
for a wrapped Keypair, I thought to start with an easy conversion from
one to the other. So you can read the keypair with the solana_sdk, then
very quickly turn it into a helium-lib keypair.

Bring the discussion!!!

* Add boosting module, function for starting hex boost

The module is named boosting so there is hopefully less confusion with
the existance of `helium_anchor_lang::hexboosting` that is autogenerated.

* merge imports

* add Io error to encode error when attempting to write a keypair

* Move ReadWrite trait to helium-lib

You can directly read a helium_lib::Keypair from a file that is a solana keypair.

* remove after bad merge

* update to use new dc_key location

* remove useless ? and Ok

* move ReadWrite back to wallet. avoid file io in helium-lib

* upgrade spl-account-compression to version compatible with anchor-lang >0.30

* move back to anchor-lang 0.29

* update helium-anchor-gen with looser dep around anchor-client

we can not go all the way back to 0.28 because of solana-sdk. So we go
back as far as we can without causing problems for upstream things like oracles.

* go back to master helium-anchor-gen that has updated anchor-lang >=0.28

* no longer writing keypairs, remove IO error variant

* cargo update after merge