Add fallback for height_for_tree
The existing height-for-tree function assumed that the entities.nft.helium.io/v2/merkles endpoint would have heights details for all trees. This -may- not be the case, and more importantly is definitely not the case on devnet.

This adds the more expensive fallback where height is computed on the fly by loading the entire merkle tree account from chain and deducing the height of the merkle tree.

Part of this required copying in definitions from the old/archived spl_account_compression crate. This allows us to just rely on the idl definition and avoiding crate dependencies